A To Do List
I made a To Do list with POO in ruby and working with multiples arquives

+++ Self Projects/To-Do List/components/main.rb
@@ -28,12 +28,21 @@ while stop == false
     to_do_list.create_a_task(title,due_date,priority,'components/tasks.txt')
   when optionAnswer == 3 # Delete a item To-Do List
     # Delete item
-    system('clear')
-    puts 'What Task do you want to delete? Say ID!'
-    puts File.open('components/tasks.txt','r').read
-    id_to_delete = gets.to_i
+    correct_Letter = false
+    while correct_Letter == false
+      sleep(1)
+      system('clear')
+      puts 'What Task do you want to delete? Say ID!'
+      puts File.open('components/tasks.txt','r').read
+      id_to_delete = gets.to_i
 
-    to_do_list.delete_a_item(id_to_delete,'components/tasks.txt')
+      if id_to_delete > 0
+        to_do_list.delete_a_item(id_to_delete,'components/tasks.txt')
+        correct_Letter = true
+      else
+        puts 'Say a correct number!'
+      end
+    end
   when optionAnswer == 4 # End
     stop = true
   else
